requirements.txt
motor==2.5.1
jinja2==3.0.3
schema==0.7.5
tornado==6.1
pygsheets==2.0.5
google-cloud-translate==3.6.1
git+https://github.com/moriW/moreover.git